FIXED: nonunique id exception when deleting vehicles

+++ src/main/java/cz/muni/fi/pa165/vozovna/service/impl/VehicleServiceImpl.java
@@ -13,6 +13,7 @@ import org.springframework.transaction.annotation.Transactional;
 
 /**
  * Implementation of Vehicle Service
+ * 
  * @author Lukas Hajek <359617@mail.muni.cz>
  */
 @Service
@@ -24,15 +25,14 @@ public class VehicleServiceImpl implements VehicleService {
     public void setUserDAO(VehicleDAO vehicleDAO) {
         this.vehicleDAO = vehicleDAO;
     }
-    
+
     @Override
-    @Transactional(readOnly=true)
+    @Transactional(readOnly = true)
     public VehicleDTO getById(Long id) {
         if (id == null) {
             return null;
         }
         Vehicle vehicle = vehicleDAO.getById(id);
-
         return new VehicleDTO(vehicle);
     }
 
@@ -42,13 +42,9 @@ public class VehicleServiceImpl implements VehicleService {
         if (vehicle == null) {
             throw new IllegalArgumentException("null vehicle");
         }
-        
         Vehicle entity = vehicle.toVehicle();
-        
         vehicleDAO.create(entity);
-        
         vehicle.fromVehicle(entity);
-        
         return entity.getId();
     }
 
@@ -58,8 +54,10 @@ public class VehicleServiceImpl implements VehicleService {
         if (vehicle == null) {
             throw new IllegalArgumentException("null vehicle");
         }
-        Vehicle v = vehicle.toVehicle();
-            
+        if (vehicle.getId() == null) {
+            throw new IllegalArgumentException("vehicle unexists");
+        }
+        Vehicle v = vehicleDAO.getById(vehicle.getId());
         vehicleDAO.remove(v);
         vehicle.setId(null);
     }
@@ -70,47 +68,54 @@ public class VehicleServiceImpl implements VehicleService {
         if (vehicle == null) {
             throw new IllegalArgumentException("null vehicle");
         }
-        
-        Vehicle entity = vehicle.toVehicle();
-        
+        if (vehicle.getId() == null) {
+            throw new IllegalArgumentException("vehicle unexists");
+        }
+        Vehicle entity = vehicleDAO.getById(vehicle.getId());
+
+        // copy DTO`s attributes into existing entity
+        entity.setBrand(vehicle.getBrand());
+        entity.setDistanceCount(vehicle.getDistanceCount());
+        entity.setEngineType(vehicle.getEngineType());
+        entity.setType(vehicle.getType());
+        entity.setVin(vehicle.getVin());
+        entity.setYearMade(vehicle.getYearMade());
+        entity.setUserClass(vehicle.getUserClass());
+
         vehicleDAO.update(entity);
-        
         vehicle.fromVehicle(entity);
-        
         return vehicle;
     }
 
     @Override
-    @Transactional(readOnly=true)
+    @Transactional(readOnly = true)
     public List<VehicleDTO> findAll() {
         List<Vehicle> vehicles = vehicleDAO.findAll();
-                
+
         // transform from List<Vehicle> to List<VehicleDTO>
         List<VehicleDTO> result = new ArrayList<>();
-        for(Vehicle vehicle:vehicles) {
+        for (Vehicle vehicle : vehicles) {
             result.add(new VehicleDTO(vehicle));
         }
-        
+
         return result;
     }
 
     @Override
-    @Transactional(readOnly=true)
+    @Transactional(readOnly = true)
     public List<VehicleDTO> findByUserClass(UserClassEnum userClass) {
         if (userClass == null) {
             throw new IllegalArgumentException("userClass");
         }
         List<Vehicle> vehicles = vehicleDAO.findByUserClass(userClass);
-                
+
         // transform from List<Vehicle> to List<VehicleDTO>
         List<VehicleDTO> result = new ArrayList<>();
-        for(Vehicle vehicle:vehicles) {
+        for (Vehicle vehicle : vehicles) {
             result.add(new VehicleDTO(vehicle));
         }
-        
+
         return result;
     }
 
-   
-    
 }
